+++ R/as_hux_table.R
@@ -26,11 +26,12 @@
 #' @family gtsummary output types
 #' @author David Hugh-Jones
 #' @examples
-#' trial %>%
-#'   dplyr::select(trt, age, grade) %>%
-#'   tbl_summary(by = trt) %>%
-#'   add_p() %>%
-#'   as_hux_table()
+#' if (requireNamespace("huxtable"))
+#'   trial %>%
+#'     dplyr::select(trt, age, grade) %>%
+#'     tbl_summary(by = trt) %>%
+#'     add_p() %>%
+#'     as_hux_table()
 #' @export
 
 as_hux_table <- function(x, include = everything(), return_calls = FALSE,
@@ -136,6 +137,26 @@ table_styling_to_huxtable_calls <- function(x, ...) {
         row = !!df_padding$row_numbers[[.x]],
         col = !!df_padding$id[[.x]],
         value = 15
+      ))
+    )
+
+  # padding2 -------------------------------------------------------------------
+  df_padding2 <-
+    x$table_styling$header %>%
+    select(.data$id, .data$column) %>%
+    inner_join(
+      x$table_styling$text_format %>%
+        filter(.data$format_type == "indent2"),
+      by = "column"
+    )
+
+  huxtable_calls[["set_left_padding2"]] <-
+    map(
+      seq_len(nrow(df_padding2)),
+      ~ expr(huxtable::set_left_padding(
+        row = !!df_padding2$row_numbers[[.x]],
+        col = !!df_padding2$id[[.x]],
+        value = 30
       ))
     )